iot-data-gen/requirement.txt
kafka-python==2.0.2
shortuuid==1.0.1
